WIP: feat: allow adding true modal dialogs to UI
Introduces the possibility to add a modal dialog to a UI in a way
that the server side will ignore interaction with all other components
(from client side) until the modal dialog is removed. Modal dialogs can
be stacked on top of each other. This is similar to setting a component disabled,
but is not the same and does not show in the UI in any way - it would be
up to the modal dialog to show a modalityy curtain on top of the outside
UI to the dialog. The feature is thus a server side security feature,
and makes the framework enforce
https://www.w3.org/TR/wai-aria-practices-1.1/#dialog_modal guidelines.

The implementation relies on a new server side only state tree feature InertData, which
makes it possible to make a component/element/statenode and its children not accept any
updates from client side, but it not the same as disabled. The inert
state is by default inherited to child nodes, except that any child can declare
itself to ignore parent inert state which means that updates are allowed for
itself and its children, unless those are explicitly made inert.

WIP:
- missing bunch of integration tests for trying to break inept UI,
which are TODOs in InertComponentIT
- the public API in the UI should be removed and only entry point would be in UIInternals.
- this also disables all UI level listeners - like UI resize listener, which might not be desirable

Resolves #10172